restructuring files, and resolve bulk actions
